pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dockingmodel"
version = "0.1.0"
description = "dimer docking"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "torch>=2.2,<2.8",
    "numpy==2.3.4",
    "scipy==1.16.2",
    "pandas==2.3.1",
    "scikit-learn==1.7.1",
    "tqdm==4.67.1",
    "lightning==2.5.2",
    "pytorch-lightning==2.5.5",
    "torchmetrics==1.8.2",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "omegaconf==2.3.0",
    "click==8.2.1",
    "python-dotenv==1.1.1",
    "einops==0.8.1",
    "mashumaro==3.16",
    "dm-tree==0.1.9",
    "biopython==1.85",
    "biotite==1.5.0",
    "biotraj==1.2.2",
    "gemmi==0.7.3",
    "OpenMM==8.3.1",
    "rdkit==2025.3.5",
    "ihm==2.7",
    "modelcif==1.4",
    "atomworks[ml]==1.0.3",
    "timm==1.0.19",
    "py3Dmol==2.5.2",
    "mediapy==1.0.2",
    "p-tqdm==1.4.2",
    "tensorboard==2.20.0",
    "wandb==0.22.3",
    "rich==14.2.0",
    "seaborn==0.13.2",
    "fire==0.7.1",
]

[project.optional-dependencies]
dev = [
    "ipykernel==6.22.0",
    "ipython==9.6.0",
    "jupyterlab==4.4.9",
    "notebook==7.4.7",
    "ipywidgets==8.1.8",
    "ipycanvas==0.14.1",
    "ipyevents==2.0.4",
    "ipdb==0.13.13",
    "ruff",
    "mypy",
    "pre-commit",
]

[project.scripts]
simplefold = "simplefold.cli:main"

[tool.setuptools.packages.find]
where = ["src", "custom"]
